大幅重构,达到论文中的程序描述 Release V1.0.0

+++ Propro-Client/Domains/Aird/WindowRange.cs
@@ -1,23 +1,21 @@
-﻿
-
-namespace AirdPro.Domains.Aird
+﻿namespace AirdPro.Domains.Aird
 {
     public class WindowRange
     {
         //前体的荷质比窗口开始位置,已经经过ExperimentDO.overlap参数调整,precursor mz
-        public float start;
+        public double start;
 
         //前体的荷质比窗口结束位置,已经经过ExperimentDO.overlap参数调整,precursor mz
-        public float end;
+        public double end;
 
         //前体的荷质比,precursor mz
-        public float mz;
+        public double mz;
 
         public string features;
 
         public WindowRange() { }
 
-        public WindowRange(float start, float end, float mz)
+        public WindowRange(double start, double end, double mz)
         {
             this.start = start;
             this.end = end;
